Move out Tag and Args from node.rs
